move execution indexer into own package
